+++ src/function/algebra/simplify.js
@@ -162,7 +162,7 @@ function factory (type, config, load, typed, math) {
   function removeParens (node) {
     return node.transform(function (node, path, parent) {
       return type.isParenthesisNode(node)
-        ? node.content
+        ? removeParens(node.content)
         : node
     })
   }
@@ -382,15 +382,13 @@ function factory (type, config, load, typed, math) {
         }
 
         // Replace placeholders with their respective nodes without traversing deeper into the replaced nodes
-        const _transform = function (node) {
+        res = res.transform(function (node) {
           if (node.isSymbolNode && matches.placeholders.hasOwnProperty(node.name)) {
             return matches.placeholders[node.name].clone()
           } else {
-            return node.map(_transform)
+            return node
           }
-        }
-
-        res = _transform(res)
+        })
 
         // const after = res.toString({parenthesis: 'all'})
         // console.log('Simplified ' + before + ' to ' + after)